+++ domain/domain-api/src/main/kotlin/by/miaskor/domain/store/repositories/ClientRepository.kt
@@ -6,7 +6,7 @@ import java.util.*
 
 interface ClientRepository : JpaRepository<ClientEntity, Int> {
   fun findByEmailOrLogin(email: String, login: String): Optional<ClientEntity>
-  fun findByBotId(botId: Int): Optional<ClientEntity>
+  fun findByBotId(botId: Long): Optional<ClientEntity>
   fun findByLogin(login: String): Optional<ClientEntity>
   fun findByEmail(email: String): Optional<ClientEntity>
 }
